% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicesFAO.R
\name{agroindexFAO}
\alias{agroindexFAO}
\title{FAO Agronomic indices}
\usage{
agroindexFAO(
  lat,
  dates,
  index.code,
  pr = NULL,
  tx = NULL,
  tn = NULL,
  pnan = 25,
  shc = 100,
  rndy = 2.5,
  rnlg = 50,
  txh = 30,
  tnh = 18
)
}
\arguments{
\item{lat}{Latitude (Integer)}

\item{dates}{Vector of dates}

\item{index.code}{Character of the index code to be computed. Options are:
"dt_st_rnagsn", "nm_flst_rnagsn", "dt_fnst_rnagsn", "dt_ed_rnagsn", "dl_agsn = rep", "dc_agsn, "rn_agsn, 
"avrn_agsn", "dc_rnlg_agsn", "tm_agsn", "dc_txh_agsn", "dc_tnh_agsn" (See Details).}

\item{pr}{Vector of precipitation data}

\item{tx}{Vector of maximum temperature data}

\item{tn}{Vector of minimum temperature data}

\item{pnan}{maximum percentage of missing data (pr, tx, tn) allowed in one year to compute the indices}

\item{shc}{maximum soil holding capacity}

\item{rndy}{amount of rain in a day to be considered as “rainy” day (in mm/day)}

\item{rnlg}{“large” rainfall event (mm/day)}

\item{txh}{“hot” tmax (degC)}

\item{tnh}{“hot” tmin (degC)}
}
\description{
Compute the selected FAO Agronomic index
}
\details{
dt_st_rnagsn = first start date of the Agronomic Season (AS)
nm_flst_rnagsn =  number of false starts of the AS
dt_fnst_rnagsn =  final start date of the AS
dt_ed_rnagsn =  end date of the AS
dl_agsn =  length of the AS
dc_agsn =  number of rainy days in AS
rn_agsn =  total amount of rain in AS
avrn_agsn =  average amount of rain in rainy days in AS
dc_rnlg_agsn =  number of days with "high" rain in AS
tm_agsn =  average daily tas in AS
dc_txh_agsn = number of "hot" tmax days in AS
dc_tnh_agsn = number of "hot" tmin days in AS
}
\author{
R. Manzanas
}
