% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicesFAO_tier1.R
\name{agroindexFAO_tier1}
\alias{agroindexFAO_tier1}
\title{Wrapper to call FAO_tier1 index calculation functions.}
\usage{
agroindexFAO_tier1(index.code, ...)
}
\arguments{
\item{index.code}{To call to the atomic function of the same name}

\item{...}{Other parameters that can be passed to the selected index function (see Details).}
}
\description{
Wrapper to call FAO_tier1 index calculation functions.
}
\details{
Index calculation is done by different functions that receive the same name 
as the corresponding index code. Therefore, type \code{?index.code} to know the arguments
that are used in each case, e.g. \code{?gsl}.
}
\section{Index codes}{

   \strong{gsl} 
   
   \strong{avg}  
   
   \strong{nd_thre}
 
   \strong{nhw}
   
   \strong{dr}
   
   \strong{prcptot}
   
   \strong{nrd}
   
   \strong{lds}
   
   \strong{sdii}
   
   \strong{prcptot_thre}
   
   \strong{ns}
}

\author{
M. Iturbide
}
