% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicesFAO.R
\name{binSpell}
\alias{binSpell}
\title{Occurrence and length of binary spells}
\usage{
binSpell(v)
}
\arguments{
\item{v}{Vector}
}
\description{
Computes ccurrence and length of binary spells)
}
\author{
R. Manzanas
}
