% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{cdd.th}
\alias{cdd.th}
\title{Cooling Degree Day (CDD)}
\usage{
cdd.th(tn, tx, tm, th = 22)
}
\arguments{
\item{tn}{Matrix or vector with minimum temperature data}

\item{tx}{Matrix or vector with maximum temperature data}

\item{tm}{Matrix or vector with mean temperature data}

\item{th}{Threshold value (Defalut is 22º)}
}
\description{
Symmetrical to the Heating Degree Day index (HDD), the CDD index is used for illustrating 
energy demand for cooling.
}
\references{
Spinoni, J., Vogt, J., and Barbosa, P. European degree-day climatologies and trends for the period 1951–2011. Int. J. Climatol. 35, 25–36. doi:10.1002/joc.3959.

Spinoni, J., Vogt, J. V, Barbosa, P., Dosio, A., McCormick, N., Bigano, A., et al. Changes of heating and cooling degree-days in Europe from 1981 to 2100. Int. J. Climatol. 38, e191–e208. doi:10.1002/joc.5362.
}
\author{
M. Iturbide
}
