% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{cddx.th}
\alias{cddx.th}
\title{Maximum length of consecutive dry days (R<1)}
\usage{
cddx.th(pr, th = 1)
}
\arguments{
\item{pr}{Vector with precipitation data}

\item{th}{Threshold below which a day is considered dry (Default is 1)}
}
\description{
Annual count of days with minimum temperature below/above a given threshold
}
\author{
M. Iturbide
}
