% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circIndexGrid.R
\name{circIndexShow}
\alias{circIndexShow}
\title{List all available circulation indices}
\usage{
circIndexShow()
}
\value{
Print a table on the screen with the following columns:
\itemize{
\item \strong{code}: Code of the index. This is the character string used as input value
for the argument \code{index.code} in \code{\link{circIndexGrid}}
\item \strong{longname}: Long description of the circulation index.
\item \strong{fun}: The name of the function used to calculate it.
\item \strong{zg, hgt, sst, slp}: A logical value (0/1) indicating the input variables required for index calculation
\item \strong{reference}: Reference for the implemented calculation.
}
}
\description{
Print a table with a summary of the available circulation indices
}
\author{
J. Bedia, M. Iturbide, A. Casanueva
}
