% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicesFAO.R
\name{computeET0}
\alias{computeET0}
\title{ET0}
\usage{
computeET0(lat, dates, tx, tn)
}
\arguments{
\item{lat}{Latitude (Integer)}

\item{dates}{Vector of dates}

\item{tx}{Vector of maximum temperature data}

\item{tn}{Vector of minimum temperature data}
}
\description{
computes ET0, accroding to Hargreaves equation
}
\details{
#' @title Occurrence and length of binary spells
#' @description Computes ccurrence and length of binary spells)
#' @param v Vector
#' @author R. Manzanas

binSpell <-  function(v) {
  ix <- c(which(v[-length(v)] != v[-1]), length(v));
  out <- list()
  out$len <- diff(c(0, ix))
  out$val <- v[ix]
  return(out)
}
}
\author{
R. Manzanas
}
