% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicesFAO_tier1.R
\name{gsl}
\alias{gsl}
\title{Function to compute the growing season length}
\usage{
gsl(tm, dates, lat, pnan = 25)
}
\arguments{
\item{tm}{Vector of mean temperature data}

\item{dates}{Matrix containing the full range of dates (ndates x 3 size); e.g. rbind(c(1995, 3, 1), c(1995, 3, 2), ...)}

\item{lat}{Latitude (Integer)}

\item{pnan}{maximum percentage of missing data (tm) allowed in one year to compute the indices}
}
\description{
see http://etccdi.pacificclimate.org/list_27_indices.shtml
}
\author{
R. Manzanas
}
