% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{hdd.th}
\alias{hdd.th}
\title{Heating Degree Days (HDD)}
\usage{
hdd.th(tn, tx, tm, th = 15.5)
}
\arguments{
\item{tn}{Matrix or vector with minimum temperature data}

\item{tx}{Matrix or vector with maximum temperature data}

\item{tm}{Matrix or vector with mean temperature data}

\item{th}{Threshold value (Defalut is 15.5º)}
}
\description{
Symmetrical to the Cooling Degree Day index (CDD), the HDD index is used for illustrating 
energy demand for heating.
}
\references{
Spinoni, J., Vogt, J., and Barbosa, P. European degree-day climatologies and trends for the period 1951–2011. Int. J. Climatol. 35, 25–36. doi:10.1002/joc.3959.

Spinoni, J., Vogt, J. V, Barbosa, P., Dosio, A., McCormick, N., Bigano, A., et al. Changes of heating and cooling degree-days in Europe from 1981 to 2100. Int. J. Climatol. 38, e191–e208. doi:10.1002/joc.5362.
}
\author{
M. Iturbide
}
