% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicesENSO.R
\name{indicesENSO}
\alias{indicesENSO}
\title{Calculation of ENSO indices from grid.}
\usage{
indicesENSO(
  grid,
  base = NULL,
  ref = NULL,
  season = NULL,
  index.code,
  members = NULL
)
}
\arguments{
\item{grid}{A grid (gridded or station dataset), or multimember grid object of sea surface temperature.}

\item{base}{Baseline grid to be substracted for the calculation of anomalies. Default: NULL. See \code{?scaleGrid}.}

\item{ref}{Reference grid to be added for the calculation of anomalies. Default: NULL. See \code{?scaleGrid}.}

\item{season}{Selected month(s) for the calculation. Default: NULL (i.e. as input grid).}

\item{index.code}{Circulation index (or vector of indices with same input variables) to be computed. See \code{circIndexShow()} for details.}

\item{members}{Select number of members. Default: NULL (all members of the grid).}
}
\value{
A list of circulation indices (and members, if applicable) with:
\itemize{
* index: vector with the time series of the teleconnection index.
* pattern: matrix with the spatial pattern of the teleconnection.
* dates and coordinates as list attributes.
}
}
\description{
Calculate ENSO indices of grids or multimember grids.
}
\details{
The calculation of ENSO indices is based on \url{https://climatedataguide.ucar.edu/climate-data/nino-sst-indices-nino-12-3-34-4-oni-and-tni}, consisting of SST anomalies, using a different size moving window.
}
\examples{
\dontrun{ 
data(ERAInterim_sst_1981_2010)
nino <- indicesENSO(grid=ERAInterim_sst_1981_2010, index.code = "NINO3.4")
}
}
