% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicesFAO_tier1.R
\name{lds}
\alias{lds}
\title{Function to compute the average/maximum length of dry (as defined by wet.threshold) spells}
\usage{
lds(
  pr,
  dates,
  length.spell = "mean",
  wet.threshold = 1,
  year = NULL,
  year.start = NULL,
  year.end = NULL,
  pnan = 25,
  lat = NULL
)
}
\arguments{
\item{pr}{Vector with daily precipitation}

\item{dates}{Matrix containing the full range of dates corresponding to "pr" (ndates x 3 size); e.g. rbind(c(1995, 3, 1), c(1995, 3, 2), ...)}

\item{length.spell}{Either "mean" or "maximum" length spell}

\item{wet.threshold}{Threshold considered to define wet days. Must be in the same units of "pr"}

\item{year}{Vector with years of interest (e.g. 1990:1995)}

\item{year.start}{Vector of dates [in "YYYY-MM-DD" format] defining the beginning of a portion of interest within each year (e.g., the agronomic season)}

\item{year.end}{Vector of dates [in "YYYY-MM-DD" format] defining the end of a portion of interest within each year (e.g., the agronomic season)}

\item{pnan}{Any year with a percentage of NA data above "pnan" will be ignored}

\item{lat}{Latitude (NULL) to indicate that latitude information is not used.}
}
\value{
Average/maximum length of dry spells (per year)
}
\description{
Function to compute the average/maximum length of dry (as defined by wet.threshold) spells
}
\author{
R. Manzanas
}
