% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{mcddd.th}
\alias{mcddd.th}
\title{Approximated Cold Degree Days}
\usage{
mcddd.th(tn, tx, th = 6)
}
\arguments{
\item{tn}{Vector with DAILY minimum temperature data}

\item{tx}{Vector with DAILY maximum temperature data}

\item{th}{Numeric (T base, default to 6).}
}
\description{
Approximation of the CDDD index (Cold Degree Days)
caluclated from daily maximum an minimum temperatures.
}
\author{
M. Iturbide
}
