% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mgdd.aux}
\alias{mgdd.aux}
\title{Threshold application for Sinusoidal Modified Growing Degree Days calculation}
\usage{
mgdd.aux(tm, th1 = 12, th2 = 18, th3 = 28, th4 = 32, last = 35)
}
\arguments{
\item{tm}{Vector with HOURLY minimum temperature data}

\item{th1}{Numeric.}

\item{th2}{Numeric.}

\item{th3}{Numeric.}

\item{th4}{Numeric.}

\item{last}{Numeric.}
}
\description{
Threshold application for Sinusoidal Modified Growing Degree Days calculation 
(see \code{mgdd.th}).
}
\details{
Defaults for the th* parameters are: th1 = 12, th2 = 18, th3 = 28, th4 = 32, last = 35.
}
\author{
M. Iturbide
}
