% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{mgdd.th}
\alias{mgdd.th}
\title{Sinusoidal Modified Growing Degree Days}
\usage{
mgdd.th(tn, tx, ...)
}
\arguments{
\item{tn}{Vector with DAILY minimum temperature data}

\item{tx}{Vector with DAILY maximum temperature data}

\item{...}{Arguments passed to function \code{mgdd.aux} (optional).
Defaults are: th1 = 12, th2 = 18, th3 = 28, th4 = 32, last = 35.}
}
\description{
Approximation of the Modified GDD index caluclated from daily maximum an minimum temperatures.
}
\author{
M. Iturbide
}
