% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearTrend.R
\name{mklr}
\alias{mklr}
\title{Error estimates of the regression coefficients}
\usage{
mklr(x, y)
}
\arguments{
\item{x}{A vector of predictors.}

\item{y}{A vector of observations.}
}
\value{
Returns a list with the estimates (coefficients: r0, r1, 
and their standard errors: sig0 and sig1).
}
\description{
Univariate linear regression with error estimates for regression coefficients.
}
\details{
The mathematics behind the calculation of the error estimates of the 
regression coefficients are based on the Dover's book by A.A.Sveshnikov 
"Problems in Probability Theory,...", p.326, --AK, Feb 27  2004.
Performs univariate linear regression of y on x:
        y = r0 + r1*x
}
\author{
J. Baño-Medina, S. Herrera
}
\keyword{internal}
