% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicesFAO_tier1.R
\name{nhw}
\alias{nhw}
\title{Function to compute the number of heatwaves (as defined by threshold and duration)}
\usage{
nhw(
  tx,
  dates,
  threshold,
  duration,
  year = NULL,
  year.start = NULL,
  year.end = NULL,
  pnan = 25
)
}
\arguments{
\item{tx}{Vector with daily maximum temperature}

\item{dates}{Matrix containing the full range of dates corresponding to "tx" (ndates x 3 size); e.g. rbind(c(1995, 3, 1), c(1995, 3, 2), ...)}

\item{threshold}{Threshold considered. Must be in the same units of "tx"}

\item{duration}{Duration (in days) considered to define the heatwave}

\item{year}{Vector with years of interest (e.g. 1990:1995)}

\item{year.start}{Vector of dates [in "YYYY-MM-DD" format] defining the beginning of a portion of interest within each year (e.g., the agronomic season)}

\item{year.end}{Vector of dates [in "YYYY-MM-DD" format] defining the end of a portion of interest within each year (e.g., the agronomic season)}

\item{pnan}{Any year with a percentage of NA data above "pnan" will be ignored}
}
\value{
Number of heatwaves (per year)
}
\description{
Function to compute the number of heatwaves (as defined by threshold and duration)
}
\author{
R. Manzanas
}
