% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{percentile}
\alias{percentile}
\title{Percentile}
\usage{
percentile(var, percent = NULL, value = NULL)
}
\arguments{
\item{var}{Vector}

\item{percent}{Number between 0 and 1 to calculate the value of a given probability.}

\item{value}{Number of the value for which the probability is going to be calculated.}
}
\description{
Calculate the value of a given probability value or the other way around.
}
\author{
M. Iturbide
}
