% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{pr.th}
\alias{pr.th}
\title{Days with precipitation below/above a given threshold}
\usage{
pr.th(pr, th = 1, operator = "<")
}
\arguments{
\item{pr}{Vector with precipitation data}

\item{th}{Threshold value (Default is 0)}

\item{operator}{Character string specifying the comparison operator (Default is "<"). 
Must be one of:
- `"<"`  (less than)
- `">"`  (greater than)
- `"<="` (less than or equal to)
- `">="` (greater than or equal to)
- `"=="` (equal to)
- `"!="` (not equal to)}
}
\description{
Annual count of days with precipitation below/above a given threshold
}
\author{
M. Iturbide
}
