% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicesFAO_tier1.R
\name{prcptot_thre}
\alias{prcptot_thre}
\title{Function to compute total amount of precipitation fallen in strong (as defined by threshold) rainy days}
\usage{
prcptot_thre(
  pr,
  dates,
  threshold = 50,
  year = NULL,
  year.start = NULL,
  year.end = NULL,
  pnan = 25
)
}
\arguments{
\item{pr}{Vector with daily precipitation}

\item{dates}{Matrix containing the full range of dates corresponding to "pr" (ndates x 3 size); e.g. rbind(c(1995, 3, 1), c(1995, 3, 2), ...)}

\item{threshold}{Threshold considered to define strong rain. Must be in the same units of "pr"}

\item{year}{Vector with years of interest (e.g. 1990:1995)}

\item{year.start}{Vector of dates [in "YYYY-MM-DD" format] defining the beginning of a portion of interest within each year (e.g., the agronomic season)}

\item{year.end}{Vector of dates [in "YYYY-MM-DD" format] defining the end of a portion of interest within each year (e.g., the agronomic season)}

\item{pnan}{Any year with a percentage of NA data above "pnan" will be ignored}
}
\value{
Total amount of precipitation fallen in strong rainy days (per year)
}
\description{
Function to compute total amount of precipitation fallen in strong (as defined by threshold) rainy days
}
\author{
R. Manzanas
}
