% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvpot.R
\name{pvpot}
\alias{pvpot}
\title{Photovoltaic potential index (pvpot)}
\usage{
pvpot(tm, wss, rad)
}
\arguments{
\item{tm}{Matrix or vector with mean temperature data (in Celsius ºC)}

\item{wss}{Matrix or vector with mean wind speed data (in meters per second m/s)}

\item{rad}{Matrix or vector with the mean radiation data (in Watios per squared meters W/m²)}
}
\value{
A numeric vector or matrix with the photovoltaic potential index, between 0-1.
}
\description{
Photovoltaic potential index describing cell potential production with respect to the optimal potential of a global downward shortwave radiation of 1000 W/m2
}
\references{
Chenni, R., M. Makhlouf, T. Kerbache, and A. Bouzid. 2007. A Detailed Modeling Method for Photovoltaic Cells. Energy 32 (9): 1724–30. https://doi.org/10.1016/j.energy.2006.12.006.

Crook, Julia A., Laura A. Jones, Piers M. Forster, and Rolf Crook. 2011. “Climate Change Impacts on Future Photovoltaic and Concentrated Solar Power Energy Output.” Energy & Environmental Science 4 (9): 3101. https://doi.org/10.1039/c1ee01495a.

Jerez, Sonia, Isabelle Tobin, Robert Vautard, Juan Pedro Montávez, Jose María López-Romero, Françoise Thais, Blanka Bartok, et al. 2015. “The Impact of Climate Change on Photovoltaic Power Generation in Europe.” Nature Communications 6 (1): 10014. https://doi.org/10.1038/ncomms10014.
}
\author{
O.Mirones
}
