% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexGrid.R
\name{read.tele}
\alias{read.tele}
\title{Read CPC teleconnections}
\usage{
read.tele()
}
\description{
Read CPC teleconnections
}
\details{
Downloaded from wget ftp://ftp.cpc.ncep.noaa.gov/wd52dg/data/indices/tele_index.nh
These are monthly tabulated indices since 1950 to present.  Indices are standardized by the 1981-2010 climatology.
}
\keyword{internal}
