% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicesFAO_tier1.R
\name{yearStartEnd}
\alias{yearStartEnd}
\title{Function to find the position marking the start and the end of a given year (or a user-defined portion of the year)}
\usage{
yearStartEnd(dates, year, year.start = NULL, year.end = NULL)
}
\arguments{
\item{dates}{Matrix containing the full range of dates (ndates x 3 size); e.g. rbind(c(1995, 3, 1), c(1995, 3, 2), ...)}

\item{year}{Year of interest (e.g. 1995)}

\item{year.start}{User-defined start of the year [in "YYYY-MM-DD" format]}

\item{year.end}{User-defined end of the year [in "YYYY-MM-DD" format]}
}
\value{
Indices marking the start and the end of a given year (or a user-defined portion of the year). The search is done within the "dates" matrix
}
\description{
Function to find the position marking the start and the end of a given year (or a user-defined portion of the year)
}
\author{
R. Manzanas
}
