% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicesFAO.R, R/indicesFAO_tier1.R
\name{binSpell}
\alias{binSpell}
\title{Occurrence and length of binary spells}
\usage{
binSpell(data)

binSpell(data)
}
\arguments{
\item{data}{Vector with data (e.g. daily precipitation)}

\item{v}{Vector}
}
\value{
Sequence of spells, including the duration of each spell. The search is done within the "data" vector
}
\description{
Computes ccurrence and length of binary spells)
}
\author{
R. Manzanas

R. Manzanas
}
