% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicesFAO_tier1.R
\name{binSpell}
\alias{binSpell}
\title{Function to compute number and length of binary (e.g. 0=dry, 1=wet) spells}
\usage{
binSpell(data)
}
\arguments{
\item{data}{Vector with data (e.g. daily precipitation)}
}
\value{
Sequence of spells, including the duration of each spell. The search is done within the "data" vector
}
\description{
Function to compute number and length of binary (e.g. 0=dry, 1=wet) spells
}
\author{
R. Manzanas
}
