% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{cddd.th}
\alias{cddd.th}
\title{Approximated Cold Degree Days}
\usage{
cddd.th(tm, th = 6)
}
\arguments{
\item{tm}{Vector with HOURLY mean temperature data}

\item{th}{Numeric (T base, default to 6).}
}
\description{
Approximation of the Modified CDDD index (Cold Degree Days)
caluclated from daily maximum an minimum temperatures.
}
\author{
M. Iturbide
}
