% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicesFAO.R
\name{computeET0}
\alias{computeET0}
\title{ET0}
\usage{
computeET0(lat, dates, tx, tn)
}
\arguments{
\item{lat}{Latitude (Integer)}

\item{dates}{Vector of dates}

\item{tx}{Vector of maximum temperature data}

\item{tn}{Vector of minimum temperature data}
}
\description{
computes ET0, accroding to Hargreaves equation
}
\author{
R. Manzanas
}
