% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{gdd.th}
\alias{gdd.th}
\title{Growing Degree Days}
\usage{
gdd.th(tm, th = 5)
}
\arguments{
\item{tm}{Vector with mean temperature data}

\item{th}{Threshold value (Defalut is 5º)}
}
\description{
Accumulated sum of the difference between daily mean temperature and the threshold (when higher than the threshold) 
over the primary growing season for mid-latitude agricultural areas in the northern Hemisphere (April-September).
}
\author{
M. Iturbide
}
