% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicesFAO_tier1.R
\name{gsl}
\alias{gsl}
\title{Function to compute the growing season length}
\usage{
gsl(tm, dates, lat, pnan = 25)
}
\arguments{
\item{tm}{Vector with mean temperature data}

\item{dates}{Matrix containing the full range of dates corresponding to "tm" (ndates x 3 size); e.g. rbind(c(1995, 3, 1), c(1995, 3, 2), ...)}

\item{lat}{Latitude (in degrees)}

\item{pnan}{Maximum percentage of missing data allowed in one year to compute the indices}
}
\value{
Number of days (per year)
}
\description{
See http://etccdi.pacificclimate.org/list_27_indices.shtml for details
}
\author{
R. Manzanas
}
