% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexGrid.R
\name{indexGrid}
\alias{indexGrid}
\title{Climate Indices in Climate4R}
\usage{
indexGrid(
  tn = NULL,
  tx = NULL,
  tm = NULL,
  pr = NULL,
  any = NULL,
  baseline = NULL,
  index.code,
  time.resolution = "year",
  ...,
  parallel = FALSE,
  max.ncores = 16,
  ncores = NULL
)
}
\arguments{
\item{tn}{A climate4R dataset of daily minimum temperature (degrees C)}

\item{tx}{A climate4R dataset of daily maximum temperature (degrees C)}

\item{tm}{A climate4R dataset of daily maximum temperature (degrees C)}

\item{pr}{A climate4R dataset of daily precipitation (mm)}

\item{any}{A climate4R dataset of any variable.}

\item{baseline}{Optional climate4R dataset. Only used if \code{index.code = "P"}, for calculating the relevant percentiles.}

\item{index.code}{Character string, indicating the specific code of the index (see Details).}

\item{time.resolution}{Output time resolution. Choices are "month", "year" (default) and "climatology".}

\item{...}{Optional. A list of arguments internally passed to the functions displayed by \code{\link{indexShow}}.}

\item{parallel}{Logical. Should parallel execution be used?}

\item{max.ncores}{Integer. Upper bound for user-defined number of cores.}

\item{ncores}{Integer number of cores used in parallel computation. Self-selected number of
cores is used when \code{ncpus = NULL} (the default), or when \code{maxcores} exceeds the default \code{ncores} value.}
}
\description{
Calculation of indices.
}
\details{
\code{\link{indexShow}} will display on screen the full list of available indices and their codes.
The names of the internal functions calculating each index is also displayed, whose help files can aid in
the definition of index-specific arguments.
}
\section{Parallel Processing}{


Parallel processing is enabled using the \pkg{parallel} package. 
Parallelization is undertaken by a FORK-type parallel socket cluster formed by \code{ncores}.
If \code{ncores} is not specified (default), \code{ncores} will be one less than the autodetected number of cores.
The maximum number of cores used for parallel processing can be set with the \code{max.ncores} argument, 
although this will be reset to the auto-detected number of cores minus 1 if this number is exceeded. Note that not all 
code, but just some critical loops within the function are parallelized.

In practice, parallelization does not always result in smaller execution times, due to the parallel overhead.
However, parallel computing may potentially provide a significant speedup for the 
particular case of large multimember datasets or large grids.
 
Parallel computing is currently not available for Windows machines.
}

\examples{
require(climate4R.datasets)
data("EOBS_Iberia_tas")
data("CFS_Iberia_tas")
fd <- indexGrid(tn = EOBS_Iberia_tas,
                time.resolution = "year",
                index.code = "FD")
per1 <- indexGrid(tn = EOBS_Iberia_tas,
                 time.resolution = "year",
                 index.code = "P",
                 percent = 90)
per2 <- indexGrid(tn = CFS_Iberia_tas,
                time.resolution = "year",
                index.code = "P",
                baseline = CFS_Iberia_tas,
                percent = 90)
hdd <- indexGrid(tn = CFS_Iberia_tas,
                 tx = CFS_Iberia_tas,
                 tm = CFS_Iberia_tas,
                 time.resolution = "year",
                 index.code = "HDD")
}
\author{
M. Iturbide
}
