% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexGrid.R
\name{indexShow}
\alias{indexShow}
\title{List all available Indices}
\usage{
indexShow()
}
\value{
Print a table on the screen with the following columns:
\itemize{
\item \strong{code}: Code of the index. This is the character string used as input value
for the argument \code{index.code} in \code{\link{indexGrid}}
\item \strong{longname}: Long description of the index
\item \strong{index.fun}: The name of the internal function used to calculate it
\item \strong{tn, tx, tm, pr}: A logical value (0/1) indicating the input variables required for index calculation
\item \strong{units}: The units of the index (when different from those of the input variable)
}
}
\description{
Print a table with a summary of the available indices
}
\author{
J. Bedia, M. Iturbide
}
