% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicesCPC.R
\name{indicesCPC}
\alias{indicesCPC}
\title{Calculation of CPC circulation indices from grid.}
\usage{
indicesCPC(grid, base, ref, season, index.code, match = match,
  n.pcs = n.pcs, rot = rot, members = members)
}
\arguments{
\item{grid}{A grid (gridded or station dataset), or multimember grid object of geopotential height or geopotential.}

\item{base}{Baseline grid to be substracted for the calculation of anomalies. Default: NULL. See \code{?scaleGrid}.}

\item{ref}{Reference grid to be added for the calculation of anomalies. Default: NULL. See \code{?scaleGrid}.}

\item{season}{Selected month(s) for the calculation. Default: NULL (i.e. as input grid).}

\item{index.code}{Circulation index (or vector of indices with same input variables) to be computed. See \code{circIndexShow()} for details.}

\item{match}{Character string with the criterion to be used for the detection of CPC indices. Options "spatial" or "temporal". Default: "spatial". See details.}

\item{n.pcs}{Integer vector with the number of EOFs to be retained for the CPC indices. Default: 10. See details.}

\item{rot}{Logical. Should VARIMAX-Rotation be performed? Default: TRUE. This argument is only relevant for CPC indices. See details.}

\item{members}{Select number of members.}
}
\value{
A list of circulation indices (and members, if applicable) with:
\itemize{
* index: vector with the time series of the teleconnection index.
* pattern: matrix with the spatial pattern of the teleconnection.
* dates and coordinates as list attributes.
* further arguments related to the CPC indices, such as the corresponding (r)EOF and (temporal or spatial, depending on \code{'match'}) correlation with the original index.
}
}
\description{
Calculate circulation indices of grids or multimember grids.
}
\details{
CPC indices are obtained, by default, as the first 10 Varimax-rotated EOFs, as explained in \url{https://www.cpc.ncep.noaa.gov/data/teledoc/telecontents.shtml}. The core of this function is \code{stats::prcomp} including Varimax rotation.
The rotated EOFs are obtained from the monthly standardized anomalies of geopotential or geopotential height at 500hPa, with a 3-month moving window.
The argument \code{match} is used to assign each rEOF to a circulation index and pattern. Matching is based on 'temporal' or 'spatial' correction of the CPC original (NCEP Reanalysis-based) indices.
}
\examples{
\dontrun{ 
data(NCEP_hgt500_2001_2010)
cpc <- indicesCPC(grid=NCEP_hgt500_2001_2010, index.code = c("NAO", "EA","PNA"), season=1)
}
}
