% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicesWT.R
\name{indicesWT}
\alias{indicesWT}
\title{Calculation of Weather types (WT) circulation indices from grid.}
\usage{
indicesWT(grid, cluster.type, centers = NULL, base = NULL, ref = NULL)
}
\arguments{
\item{grid}{A grid (gridded or station dataset), or multimember grid object.}

\item{cluster.type}{Weather type clustering method. See details.}

\item{centers}{Integer value indicating the number of clusters, \strong{k}, or center points. See details.}

\item{base}{Baseline grid to be substracted for the calculation of anomalies. Default: NULL. See \code{?scaleGrid}.}

\item{ref}{Reference grid to be added for the calculation of anomalies. Default: NULL. See \code{?scaleGrid}.}
}
\value{
The WT circulation indices (and members, if applicable) with:
\itemize{
\item index: vector with the corresponding cluster from each point of the series.
\item pattern: Array with the spatial pattern of the clusters obtained.
\item dates and coordinates.
\item further arguments related to the WT indices, such as the corresponding cluster type, number of clusters, etc.
}
}
\description{
Calculate circulation indices of grids or multimember grids.
}
\details{
The clustering parameters for weather typing are internally passed to \code{\link[transformeR]{clusterGrid}}.
The function calculates the weather types from the season especified as a whole.
}
\examples{
\dontrun{
data(NCEP_slp_2001_2010)
wt <- indicesWT(grid=NCEP_slp_2001_2010, cluster.type="kmeans", centers = 3)
}
}
