% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicesWT.R
\name{indicesWT}
\alias{indicesWT}
\title{Calculation of Weather types (WT) circulation indices from grid.}
\usage{
indicesWT(grid, season, cluster.type, centers = centers)
}
\arguments{
\item{grid}{A grid (gridded or station dataset), or multimember grid object of geopotential height or geopotential.}

\item{season}{Selected month(s) for the calculation. Default: NULL (i.e. as input grid).}

\item{cluster.type}{Clustering type that matches with the WT circulation index to be computed.}

\item{centers}{Integer value indicating the number of clusters, \strong{k}, or center points. See \link[TransformeR]{clusterGrid} for more details.}
}
\value{
The WT circulation indices (and members, if applicable) with:
\itemize{
* index: vector with the corresponding cluster from each point of the series.
* pattern: Array with the spatial pattern of the clusters obtained.
* dates and coordinates.
* further arguments related to the WT indices, such as the corresponding cluster type, number of clusters, etc.
}
}
\description{
Calculate circulation indices of grids or multimember grids.
}
\examples{
\dontrun{ 
data(NCEP_hgt500_2001_2010)
wt <- indicesWT(grid=NCEP_hgt500_2001_2010, season=1, season=season, cluster.type="kmeans", centers=10)
}
}
