% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambWT.R
\name{lambWT}
\alias{lambWT}
\title{Calculation of Lamb Weather types (WT).}
\usage{
lambWT(grid, center.point, season, base, ref)
}
\arguments{
\item{grid}{A grid (gridded or station dataset), or multimember grid object of MSLP values.}

\item{center.point}{A two value vector that must include lon and lat from a location that will work as center point for the Lamb WT.
See details.}

\item{season}{Selected month(s) for the calculation. Default: NULL (i.e. as input grid).}

\item{base}{Baseline grid to be substracted for the calculation of anomalies. Default: NULL. See \code{?scaleGrid}.}

\item{ref}{Reference grid to be added for the calculation of anomalies. Default: NULL. See \code{?scaleGrid}.}
}
\value{
wtseries = column vector of discrete weather types defined as follows:

purely anticyclonic = 1
directional anticyclonic from NE to N = 2 to 9
purely directional from NE to N = 10 to 17
purely cyclonic = 18
directional cyclonic from NE to N = 19 to 26
light indeterminate flow N = 27
}
\description{
Calculate automated Lamb WT as defined in Trigo daCamara 2000 and Jones et al. 2012, 
both Int J Climatol
}
\details{
According to Jones et al. 2012 (Int J Climatol), Lamb WT is only applied on North Atlantic domain. 
The input grid units must be Pa, not hPa/mbar. If it is not in Pa, the units will be converted automatically.
A center location point must be specified by the user. Then, the function calculates the from left to right and from first to 16st 
the rest of the location point from the grid specified by Jones et al. 2012:
 
 
          01    02
    03    04    05    06
    07    08    09    10
    11    12    13    14
          15    16


where the north-south distance is 5º and the west-east distance is 10º.
}
\examples{

}
