% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvpot.R
\name{pvpot}
\alias{pvpot}
\title{Photovoltaic potential index (pvpot)}
\usage{
pvpot(tm, wss, rad, clamp = FALSE)
}
\arguments{
\item{tm}{Numeric vector, array or matrix with mean air temperature in °C.}

\item{wss}{Numeric vector, array or matrix with mean wind speed in m s^-1.}

\item{rad}{Numeric vector, array or matrix with mean surface solar radiation in
W m^-2. If radiation is provided as accumulated energy in J m^-2, it must be
converted beforehand by dividing by the accumulation period in seconds
(e.g. 86400 for daily data, 3600 for hourly data).}

\item{clamp}{Logical. If TRUE, clamp output to the [0, 1] interval.}
}
\value{
Numeric vector, array or matrix with the photovoltaic potential index.
}
\description{
Photovoltaic potential index describing cell potential production
with respect to the optimal potential at a global downward shortwave radiation
of 1000 W m^-2.
}
