% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{tn.th}
\alias{tn.th}
\title{Days with minimum temperature below a given threshold}
\usage{
tn.th(tn, th = 0)
}
\arguments{
\item{tn}{Vector with minimum temperature data}

\item{th}{Threshold value (Default is 0)}
}
\description{
Annual count of days with maximum temperature above a given threshold
}
\author{
M. Iturbide
}
