% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{tx.th}
\alias{tx.th}
\title{Days with maximum temperature above a given threshold}
\usage{
tx.th(tx, th)
}
\arguments{
\item{tx}{Vector with minimum temperature data}

\item{th}{Threshold value}
}
\description{
Annual count of days with maximum temperature above a given threshold
}
\author{
M. Iturbide
}
